cloud/requirements.txt
beautifulsoup4==4.14.3
boto3==1.42.24
botocore==1.42.24
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
jmespath==1.0.1
lxml==6.0.2
numpy==2.2.6
pandas==2.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
soupsieve==2.8.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
